+++ Examen_IB/app/src/main/java/com/example/examen_ib/dtos/SucursalDto.kt
@@ -6,7 +6,7 @@ class SucursalDto(
     val ciudad:String,
     val direccion: String,
     val servicioTecnico: Boolean,
-    val numeroEmpleados: Int,
+    val numeroEmpleados: Long,
     val fechaApertura: String,
     val supermercadoId: String
 ) {
